updated to the working one

+++ src/wgpugst/qtpreceive/sendinput.rs
@@ -7,14 +7,14 @@ pub async fn send_mouse<'a>(
     mouse_state: u32,
     cursor_position: &'a PhysicalPosition<f64>,
     sent_stream_input: &'a Connection,
-    mouse_data: u32,
+    mouse_data: i8,
 ) {
     let mut builder = Builder::default();
     let mut send_input = builder.start_map();
     send_input.push("mouse_state", mouse_state as u32);
     send_input.push("position_x", cursor_position.x as i16);
     send_input.push("position_y", cursor_position.y as i16);
-    send_input.push("mouse_data", mouse_data as u8);
+    send_input.push("mouse_data", mouse_data as i8);
     send_input.end_map();
     // println!(
     //     "Mouse state :{} x :{} y :{}",
